remove in memory test framework
