--- test/snapshot/test_runner_e2e.go
@@ -1,1 +1,0 @@
-package snapshot
